update to v2.9007.486 version and update documentation filename set to match what is delivered with that build.
create documentation shortcuts that start via notepad, to accomodate
text files with non-standard extensions.

swap order of remove/repair radio buttons so that remove is the first choice.

add CustomActions in native DLL to workaround custom action type=34 with
msidbCustomActionTypeContinue does not seem to reliably ignore all errors.

move execution of client shutdown sequences to be before InstallValidate
so that files currently in use will not trigger a warning dialog.